+++ client/launcher/build.sh
