+++ client/launcher/src/main/resources/logback.xml
