+++ client/launcher/src/main/resources/reference.conf
