+++ client/src/main/scala/raw/client/utils/AuthenticatedUser.scala
@@ -10,7 +10,7 @@
  * licenses/APL.txt.
  */
 
-package raw.utils
+package raw.client.utils
 
 import com.fasterxml.jackson.annotation.JsonSubTypes.{Type => JsonType}
 import com.fasterxml.jackson.annotation.{JsonSubTypes, JsonTypeInfo}
