+++ client/src/main/scala/raw/client/utils/RawConcurrentHashMap.scala
@@ -10,7 +10,7 @@
  * licenses/APL.txt.
  */
 
-package raw.utils
+package raw.client.utils
 
 import java.util.concurrent.{ConcurrentHashMap, ExecutionException}
