+++ client/src/main/scala/raw/client/utils/RawService.scala
@@ -10,7 +10,7 @@
  * licenses/APL.txt.
  */
 
-package raw.utils
+package raw.client.utils
 
 import com.typesafe.scalalogging.StrictLogging
