+++ client/src/main/scala/raw/client/utils/Uid.scala
@@ -10,11 +10,8 @@
  * licenses/APL.txt.
  */
 
-package raw.utils
+package raw.client.utils
 
-/**
- * Top-level Exception.
- * Message contains information that can be shared with the end-user.
- * TODO (msb): Add methods formalizing error codes, etc.
- */
-class RawException(message: String, cause: Throwable = null) extends RuntimeException(message, cause)
+final case class Uid(uid: String) extends AnyVal {
+  override def toString: String = uid
+}
