+++ client/src/test/scala/raw/utils/RawTestSuite.scala
@@ -10,7 +10,7 @@
  * licenses/APL.txt.
  */
 
-package raw.utils
+package raw.client.utils
 
 import com.typesafe.scalalogging.StrictLogging
 import org.scalatest.funsuite.FixtureAnyFunSuite
