+++ client/src/test/scala/raw/utils/Tags.scala
@@ -10,7 +10,7 @@
  * licenses/APL.txt.
  */
 
-package raw.utils
+package raw.client.utils
 
 import org.scalatest.Tag
