+++ client/src/test/scala/raw/utils/TrainingWheelsContext.scala
@@ -10,7 +10,7 @@
  * licenses/APL.txt.
  */
 
-package raw.utils
+package raw.client.utils
 
 trait TrainingWheelsContext {
   this: RawTestSuite with SettingsTestContext =>
