--- utils/.sbtopts
@@ -1,5 +1,0 @@
--J-Xmx4G
--J-XX:+UseG1GC
--J--add-opens=java.base/java.lang=ALL-UNNAMED
--J--add-opens=java.base/java.util=ALL-UNNAMED
--J-XX:+CrashOnOutOfMemoryError
