--- utils/.scalafmt.conf
@@ -1,138 +1,0 @@
-version = 2.7.5
-maxColumn = 120
-
-project.excludeFilters = ["/inkytonik/.*"]
-
-optIn.forceBlankLineBeforeDocstring = true
-trailingCommas = "never"
-
-align {
-  openParenCallSite = false
-  openParenDefnSite = false
-  preset=some
-  stripMargin = false
-  tokens = []
-}
-
-assumeStandardLibraryStripMargin = true
-binPack.parentConstructors = "never"
-
-continuationIndent {
-  callSite = 2
-  ctorSite = 4
-  defnSite = 4
-  extendSite = 4
-}
-
-danglingParentheses {
-  ctrlSite = true
-  defnSite = true
-}
-
-
-docstrings.style = "Asterisk" // JavaDoc
-includeCurlyBraceInSelectChains = true
-includeNoParensInSelectChains = false
-
-literals {
-  double = "Lower"
-  float = "Lower"
-  hexDigits = "Lower"
-  hexPrefix = "Lower"
-  long = "Upper"
-}
-
-newlines {
-  afterCurlyLambdaParams = squash
-  afterCurlyLambdaParams=squash
-  alwaysBeforeElseAfterCurlyIf = false
-  beforeCurlyLambdaParams = multilineWithCaseOnly
-  beforeMultiline = fold
-  beforeMultilineDef = false
-  implicitParamListModifierPrefer = before
-}
-
-rewrite {
-  redundantBraces {
-    generalExpressions = false
-    ifElseExpressions = false
-    methodBodies = false
-    parensForOneLineApply = true
-    stringInterpolation = true
-  }
-  rules = [
-    RedundantBraces,
-    RedundantParens,
-    SortModifiers,
-    SortImports
-    ]
-  redundantBraces {
-    generalExpressions = false
-    methodBodies = false
-    ifElseExpressions = false
-    parensForOneLineApply = true
-    stringInterpolation = true
-  }
-  sortModifiers {
-    order = [ 
-      "implicit",
-      "final",
-      "sealed",
-      "abstract",
-      "override",
-      "private",
-      "protected",
-      "lazy"
-    ]
-  }
-}
-
-continuationIndent {
-  callSite = 2
-  defnSite = 4
-  ctorSite = 4
-  extendSite = 4
-}
-
-align {
-  preset=some
-  tokens = []
-  openParenCallSite = false
-  openParenDefnSite = false
-  preset = some
-  stripMargin = false
-  tokens = []
-}
-
-danglingParentheses {
-  defnSite = true
-  ctrlSite = true
-}
-
-spaces {
-  afterKeywordBeforeParen = true
-  afterSymbolicDefs = false
-  beforeContextBoundColon = "Never"
-  inByNameTypes = true
-  inParentheses = false
-  neverAroundInfixTypes = []
-  afterKeywordBeforeParen = true
-  inByNameTypes = true
-  afterSymbolicDefs = false
-}
-
-literals {
-  long = "Upper"
-  float = "Lower"
-  double = "Lower"
-  hexPrefix = "Lower"
-  hexDigits = "Lower"
-}
-
-docstrings {
-  style = "Asterisk" // JavaDoc
-}
-
-binPack {
-  parentConstructors = "Never"
-}
