--- utils/project/build.properties
@@ -1,1 +1,0 @@
-sbt.version=1.9.6
